Safari-611.1.10/LayoutTests/platform/mac-mojave/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-method.any.worker-expected.txt


PASS Response.redirected should be false on not-redirected responses
PASS Redirect 301 with GET
PASS Redirect 301 with POST
FAIL Redirect 301 with HEAD assert_equals: Request Content-Length after redirection is NO expected "NO" but got "0"
PASS Redirect 302 with GET
PASS Redirect 302 with POST
FAIL Redirect 302 with HEAD assert_equals: Request Content-Length after redirection is NO expected "NO" but got "0"
PASS Redirect 303 with GET
PASS Redirect 303 with POST
FAIL Redirect 303 with HEAD assert_equals: Request Content-Length after redirection is NO expected "NO" but got "0"
PASS Redirect 303 with TESTING
PASS Redirect 307 with GET
PASS Redirect 307 with POST (string body)
FAIL Redirect 307 with POST (blob body) assert_equals: Request Content-Type after redirection is NO expected "NO" but got "application/x-www-form-urlencoded"
FAIL Redirect 307 with HEAD assert_equals: Request Content-Length after redirection is NO expected "NO" but got "0"
